requirements.txt
numpy
matplotlib
pickle
tqdm
pytorch
torchvision
pytorch-ignite
scikit-learn
